Update imageSubmitter.html
Added comments throughout the document for ease of use when combining it into a dashboard later. 

Changed Date Displayed with Images to removed second count to improve readability of the USER PHOTO DIARY.

Fixed styling to better match theme of current login page. 
Although, this document will not have css styling on the same file as it later down the line.