Create graphDemo.html
Setup basic HTML template to be used to display plant growth data graphically to user. Minimal Styling done so demo is presentable. Will later be combined with other components in a dashboard so no more time will be spent on styling.

Yet to implement graphing functionality. 
Most likely to be implemented with JS and html tools that would be appropriate. (<canvas> seems to be a good way to draw graphs)