drawGraph() added!
Finally visualise the inputs the user gives our site. 

Padding found by looking at the max values collected for the given plant name selected. 
Data will be plotted for each data point in the object.
Then titles added to the axis (this needs to be done last as it may mess with the formatting of the graph)